feat : Fixing data types  and data investigation
String data types are converted and data is being is investigated for
abnormalites

All the data in the csv files are in string format.This string format is
changed to respective format.
There are less number of students in the diaily engagement than the
enrolled students.This problem is being investigated.